% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implement_wetland.R
\name{add_connection}
\alias{add_connection}
\title{Add a connection to a connectivity input table.}
\usage{
add_connection(
  obj_con,
  id_i,
  obj_typ_add,
  obj_id_add,
  hyd_typ_add,
  frac_add,
  position = Inf
)
}
\arguments{
\item{obj_con}{The object connectivity input table}

\item{obj_typ_add}{Object type of the newly added connection}

\item{obj_id_add}{Object id into which the newly added connection is routed.}

\item{hyd_typ_add}{Type of hydrological flux of the newly added connection.}

\item{frac_add}{Flow fraction of the newly added connection (will be
normalized to 1 in case the fractions sum with existing fluxes do not add
up to 1).}

\item{position}{Position in the input table where the new connection is
added.}
}
\value{
The updated connectivity input table.
}
\description{
Add a connection to a connectivity input table.
}
\keyword{internal}
