% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_landuse.R
\name{add_dtl_op}
\alias{add_dtl_op}
\title{Add a landuse decision table operation to a management schedule}
\usage{
add_dtl_op(swat_inputs, hru_id, op_names)
}
\arguments{
\item{swat_inputs}{List with SWAT+ input files.}

\item{hru_id}{HRU IDs for which the land use is updated.}

\item{op_names}{Text string which indicates the names of the decision table
operations}
}
\value{
The SWAT+ input tables list with the updated management.sch table.
}
\description{
Add a landuse decision table operation to a management schedule
}
\keyword{internal}
