% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_lead_zeros}
\alias{add_lead_zeros}
\title{Convert number to character and add leading zeros based on existing IDs}
\usage{
add_lead_zeros(i, ids)
}
\arguments{
\item{i}{ID i which is converted}

\item{ids}{Numeric vector of existing IDs}
}
\value{
The value of i as character with leading zeros
}
\description{
Convert number to character and add leading zeros based on existing IDs
}
\keyword{internal}
