% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write_inputs.R
\name{as_mtx_null}
\alias{as_mtx_null}
\title{Transform x to a matrix with 7 columns and fill up with NA values}
\usage{
as_mtx_null(x, n)
}
\arguments{
\item{x}{character vector or NULL}

\item{n}{Number of elements}
}
\description{
Transform x to a matrix with 7 columns and fill up with NA values
}
\keyword{internal}
