% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implement_pond.R
\name{check_arguments_pond}
\alias{check_arguments_pond}
\title{Check the function input arguments.}
\usage{
check_arguments_pond(swat_inputs, hru_id, to_cha_id, from_cha_id)
}
\arguments{
\item{swat_inputs}{List with SWAT+ input files.}

\item{hru_id}{HRU IDs which will be replaced by ponds.}

\item{to_cha_id}{Channel IDs into which implemented ponds route water.}

\item{from_cha_id}{(optional) channel IDs which send water to the implemented
ponds.}
}
\value{
A boolean vector to indicate which HRUs are already replaced by
ponds.
}
\description{
Lengths and data types are checked. Routine checks if HRUs were already
replaced by ponds, triggers a warning if yes and removes the HRUs from the
replacement set.
}
\keyword{internal}
