% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implement_wetland.R
\name{check_is_wetland}
\alias{check_is_wetland}
\title{Check if HRUs are already wetlands..}
\usage{
check_is_wetland(swat_inputs, hru_id)
}
\arguments{
\item{swat_inputs}{List with SWAT+ input files.}

\item{hru_id}{HRU IDs to which surface water storage (wetland) will be added.}
}
\value{
A boolean vector to indicate which HRUs are already wetlands.
}
\description{
Check if HRUs are already wetlands..
}
\keyword{internal}
