% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_nswrm_definition.R
\name{check_settings_column}
\alias{check_settings_column}
\title{Check columns and types of data for settings input table.}
\usage{
check_settings_column(tbl, col_name, val_class, measr_type)
}
\arguments{
\item{tbl}{Settings table.}

\item{col_name}{Name of column which is checked and/or coerced.}

\item{val_class}{Class which data should have (e.g. character, integer, or numeric).}

\item{measr_type}{Type of measure, either pond, wetland or all.}
}
\value{
The input table with data type coerced to val_class if necessary.
}
\description{
Check columns and types of data for settings input table.
}
\keyword{internal}
