% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_management.R
\name{compare_scen_quo_properties}
\alias{compare_scen_quo_properties}
\title{Compare start/end years and numbers of HRUs for all scenarios with the status
quo.}
\usage{
compare_scen_quo_properties(farm_scen, farm_squo)
}
\arguments{
\item{farmr_scen}{List of scenario farmR projects.}

\item{farmr_quo}{farmR project for the status quo case.}
}
\value{
Triggeres errors if differences are identified.
}
\description{
Compare start/end years and numbers of HRUs for all scenarios with the status
quo.
}
\keyword{internal}
