% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_management.R
\name{date_to_monday}
\alias{date_to_monday}
\title{Convert dates to numeric month and day values}
\usage{
date_to_monday(tbl)
}
\arguments{
\item{tbl}{Operation schedule table.}
}
\value{
The table without date column and updated month and day columns.
}
\description{
Convert dates to numeric month and day values
}
\keyword{internal}
