% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implement_pond.R, R/implement_wetland.R
\name{get_chg_ids_pond}
\alias{get_chg_ids_pond}
\title{Get the IDs of the units for which the connectivities must be changed.}
\usage{
get_chg_ids_pond(con_tbl, hru_ids)

get_chg_ids_pond(con_tbl, hru_ids)
}
\arguments{
\item{con_tbl}{A SWAT+ .con input table (connectivity)}

\item{hru_ids}{Numeric vector of HRU IDs which will be replaced by ponds.}
}
\value{
A tibble with the objects for which the connectivities must be
changed. The table has the following columns:
\itemize{
\item \code{obj_id} ID of the object for which the connectivity must be changes
\item \code{id} ID of the HRU which will be replaced by pond
\item \code{con_id} ID of the connectivity in the connectivity table which must be
changed.
}

A tibble with the objects for which the connectivities must be
changed. The table has the following columns:
\itemize{
\item \code{obj_id} ID of the object for which the connectivity must be changes
\item \code{id} ID of the HRU which will be replaced by pond
\item \code{con_id} ID of the connectivity in the connectivity table which must be
changed.
}
}
\description{
The units in the connectivity input table are identified which initially
routed water into the HRUs which will be replaced by ponds.

The units in the connectivity input table are identified which initially
routed water into the HRUs which will be replaced by ponds.
}
\keyword{internal}
