% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_nswrm_definition.R
\name{get_mgt_short_labels}
\alias{get_mgt_short_labels}
\title{Derive short labels for the management names to be used as suffixes in the
management definition files.}
\usage{
get_mgt_short_labels(mgt_labels)
}
\arguments{
\item{mgt_lables}{Vector with management scenario names.}
}
\value{
A vector with short labels with 3 characters.
}
\description{
Derive short labels for the management names to be used as suffixes in the
management definition files.
}
\keyword{internal}
