% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implement_nswrm.R
\name{implement_nswrm}
\alias{implement_nswrm}
\title{Implement NSWRMs in the SWAT+ model input tables.}
\usage{
implement_nswrm(nswrm_id, nswrm_defs, swat_inputs)
}
\arguments{
\item{nswrm_defs}{List with already loaded NSWRM definition tables}

\item{swat_inputs}{List with SWAT+ input files.}

\item{file_path}{Path to the '.csv' location file.}
}
\value{
The list of SWAT+ input tables with NSWRMs implemented in the
respective input tables.
}
\description{
Implement NSWRMs in the SWAT+ model input tables.
}
\keyword{internal}
