% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implement_pond.R
\name{implement_ponds}
\alias{implement_ponds}
\title{Implement ponds by replacing land objects.}
\usage{
implement_ponds(swat_inputs, hru_to_pond)
}
\arguments{
\item{swat_inputs}{List with SWAT+ input files.}

\item{hru_to_pond}{Tibble which defines the changes from land objects to pond
objects.}
}
\value{
The \code{swat_inputs} list with updated input tables which include all
necessary changes to replace land objects by ponds.
}
\description{
\code{implement_ponds()} uses the \code{hru_to_pond} input table to implement ponds.
The land objects defined by the \code{hru_id} in \code{hru_to_pond} are assigned an
area of (almost) 0. The connection fractions for those land objects are
set to 0. Reservoirs with the same area/lat/long/elev are generated (at the
moment with default reservoir dimensions). All land objects which initially
routed water to the land objects with \code{hru_id} now route water to the new
reservoirs. The new reservoirs must route water to one channel each defined
with \code{to_cha_id} in \code{hru_to_pond}. Optionally a \code{from_cha_id} can be defined
if channels should be rerouted into the new reservoir.
}
