% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implement_wetland.R
\name{is_hru_drained}
\alias{is_hru_drained}
\title{Check if an HRU has tile drainage implemented or not.}
\usage{
is_hru_drained(swat_inputs, hru_id)
}
\arguments{
\item{swat_inputs}{List with SWAT+ input files.}

\item{hru_id}{HRU IDs to which surface water storage (wetland) will be added.}
}
\value{
A boolean vector to indicate which HRUs are drained.
}
\description{
Check if an HRU has tile drainage implemented or not.
}
\keyword{internal}
