% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implement_nswrm.R
\name{link_hru_res_ids}
\alias{link_hru_res_ids}
\title{Link the HRU IDs which were replaced by a pond or a constructed wetland with
the IDs of the new objects in the reservoir.res input file.}
\usage{
link_hru_res_ids(hru_ids, res_res, type)
}
\arguments{
\item{hru_ids}{List of vectors of HRU IDs}

\item{res_res}{Updated reservoir.res SWAT+ input table}

\item{type}{Type of measures (either \code{'pnd'} or \code{'cwl'})}
}
\value{
A list with the vectors of \code{hru_ids} and corresponding \code{res_ids}.
The vectors have the same length,
}
\description{
Link the HRU IDs which were replaced by a pond or a constructed wetland with
the IDs of the new objects in the reservoir.res input file.
}
\keyword{internal}
