% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_load_project.R
\name{load_measr}
\alias{load_measr}
\title{Load an existing SWATmeasR project}
\usage{
load_measr(file)
}
\arguments{
\item{file}{File path of a SWATmeasR project located in a SWAT+ project
folder.}
}
\value{
Loads an existing SWATmeasR project (as an R6 object) into the R
working environment and saves the project the SWAT+ project folder.
}
\description{
Load an existing SWATmeasR project
}
