% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_nswrm_definition.R
\name{load_mgt_def}
\alias{load_mgt_def}
\title{Load the definition input table for NSWRMs which are represented by a land
use change}
\usage{
load_mgt_def(file_path, swat_inputs)
}
\arguments{
\item{file_path}{Path to the '.csv' definition file.}

\item{swat_inputs}{List with SWAT+ input files.}
}
\value{
The loaded land use definition table as a tibble.
}
\description{
Load the definition input table for NSWRMs which are represented by a land
use change
}
\keyword{internal}
