% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_nswrm_definition.R
\name{load_nswrm_def}
\alias{load_nswrm_def}
\title{Load a definition table for an NSWRM type from a '.csv' file into the
\code{measr_project}.}
\usage{
load_nswrm_def(file_path, type, nswrm_defs, swat_inputs, overwrite)
}
\arguments{
\item{file_path}{Path to the '.csv' or '.rds' definition file.}

\item{type}{Type of the NSWRM which is defined by this input file. The
type must be one of the options \code{'land_use'}, \code{'management'},\code{'pond'},
\code{'wetland'}, or \code{'constr_wetland'}.}

\item{nswrm_defs}{List with already loaded NSWRM definition tables}

\item{swat_inputs}{List with SWAT+ input files.}

\item{overwrite}{Overwrite existing definition table? If \code{TRUE} existing
definition table can be overwritten.}
}
\value{
The list with NSWRM definition tables with the new added entry.
}
\description{
Load a definition table for an NSWRM type from a '.csv' file into the
\code{measr_project}.
}
\details{
The different \code{type} options represent different groups of NSWRMs. The
different options group the following measures:
\itemize{
\item \code{'land_use'}: A land use measure definition table can include the
definitions for all NSWRMs which are represented by a land use
change. The lookup table can include the settings for the NSWRM codes
\strong{buffer}, \strong{edgefilter}, \strong{hedge}, \strong{grassslope}, \strong{grassrchrg}
, and \strong{afforest}. The \code{land_use} table must provide the columns
\code{type}, \code{plnt_com}, \code{mgt}, \code{cn2}, \code{cons_prac}, and
\code{ov_mann}.
\item \code{'management'}: The management definition is an \verb{*.rds} object which
must be prepared with the function
\code{prepare_management_scenario_inputs()}. The function uses the
\code{SWATfarmR} projects which are located in the \code{project_path} and
generates the file \code{'mgt_scenarios.rds'} which is the required input
to define management related NSWRMs.
\item \code{'pond'}: A pond definition table includes all definitions for pond
locations. The \code{pond} table must provide the columns \code{hru_id},
\code{to_cha_id}, and \code{from_cha_id}.
Optionally, the hydrology.res parameters for a pond and the pointers
for \code{'rel'} (release decision table), \code{'sed'} for sediment.res
parametrisations, and \code{'nut'} for nutrients.res parametrisations can
be defined.
\item \code{'constr_wetland'}: A constructed wetland is implemented the same way
as a pond. Thus the same definitions like for the pond apply for a
constructed wetland. As a constructed wetland is considered to be part
of the channel/reservoir network, the definitions of \code{to_cha_id} and
\code{from_cha_id} are mandatory.
\item \code{'wetland'}: A wetland definition table includes all definitions for
wetland locations. The \code{wetland} table must provide the columns \code{hru_id}.
Optionally, water from a wetland can be directly routed into a channel
and overwrite the original routing of the wetland HRU. In that case
\code{to_cha_id} must be defined. The hydrology.wet parameters for a
wetland and the pointers for \code{'rel'} (release decision table),
\code{'sed'} for sediment.res parametrisations, and  \code{'nut'} for
nutrients.res parametrisations can be defined.
}
}
\keyword{internal}
