% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_nswrm_definition.R
\name{load_nswrm_loc}
\alias{load_nswrm_loc}
\title{Load the locations of NSWRMs (links to spatial objects) from a '.csv' file
into the \code{measr_project}.}
\usage{
load_nswrm_loc(file_path, nswrm_defs, swat_inputs, overwrite)
}
\arguments{
\item{file_path}{Path to the '.csv' location file.}

\item{nswrm_defs}{List with already loaded NSWRM definition tables}

\item{swat_inputs}{List with SWAT+ input files.}

\item{overwrite}{Overwrite existing location table? If \code{TRUE} existing
location table can be overwritten.}
}
\value{
The list with NSWRM definition tables with the NSWRM location table
added.
}
\description{
Load the locations of NSWRMs (links to spatial objects) from a '.csv' file
into the \code{measr_project}.
}
\keyword{internal}
