% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_nswrm_definition.R
\name{load_pond_def}
\alias{load_pond_def}
\title{Load the definition input table for HRUs which will be replaced by ponds}
\usage{
load_pond_def(file_path, swat_inputs)
}
\arguments{
\item{file_path}{Path to the '.csv' definition file.}

\item{swat_inputs}{List with SWAT+ input files.}
}
\value{
The loaded pond definition table as a tibble.
}
\description{
Load the definition input table for HRUs which will be replaced by ponds
}
\keyword{internal}
