% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_nswrm_definition.R
\name{load_water_def}
\alias{load_water_def}
\title{Load the definition input table for HRUs which will be replaced by ponds,
or where wetlands will be added}
\usage{
load_water_def(file_path, swat_inputs, type)
}
\arguments{
\item{file_path}{Path to the '.csv' definition file.}

\item{swat_inputs}{List with SWAT+ input files.}

\item{type}{Either 'pond' or 'wetland'. Label controls different settings in
definition}
}
\value{
The loaded pond/wetland definition table as a tibble.
}
\description{
Load the definition input table for HRUs which will be replaced by ponds,
or where wetlands will be added
}
\keyword{internal}
