% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{match_ids}
\alias{match_ids}
\title{Find the same sequence of ids in a vector compared to vectors in a list.}
\usage{
match_ids(id_list, id_vct)
}
\arguments{
\item{id_list}{List of id vectors}

\item{id_vct}{Vector of ds}
}
\value{
A boolean vector which shows the matching id sequences.
}
\description{
Find the same sequence of ids in a vector compared to vectors in a list.
}
\keyword{internal}
