% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_load_project.R
\name{new_measr}
\alias{new_measr}
\title{Initiate a new SWATmeasR project}
\usage{
new_measr(project_name, project_path)
}
\arguments{
\item{project_name}{the name of the new \code{measr_project}.}

\item{project_path}{Path to the SWAT TxtInOut folder to which the farmr
should be applied.}
}
\value{
Generates a new measr_project in the working environment (as an R6
object) and saves the project the SWAT+ project folder.
}
\description{
Initiate a new SWATmeasR project
}
