% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_inputs.R
\name{paste_expression}
\alias{paste_expression}
\title{Build vector expression for parsing}
\usage{
paste_expression(x)
}
\arguments{
\item{x}{character vector.}
}
\value{
A character string of structure c(...).
}
\description{
Build vector expression for parsing
}
\keyword{internal}
