% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_management.R
\name{prepare_management_scenario_inputs}
\alias{prepare_management_scenario_inputs}
\title{Prepare the management input data from SWATfarmR project as inputs for the
definition of management related NSWRMs with
\code{measr_project$load_nswrm_definition()}.}
\usage{
prepare_management_scenario_inputs(
  project_path,
  status_quo,
  scenarios = NULL,
  synonyms = NULL,
  start_year = NULL,
  end_year = NULL,
  write_path = project_path,
  write_csv_mgts = FALSE,
  update_luse_label = FALSE,
  sync_dates = TRUE
)
}
\arguments{
\item{project_path}{Path to the SWAT project folder on the hard drive (i.e.
txtinout folder), or a folder with sub-folders for the status quo and the
scenarios, where each folder contains the files hru-data.hru, landuse.lum,
management.sch, and plant.ini for each case.}

\item{status_quo}{Name of the SWATfarmR project which provides the status
quo.}

\item{scenarios}{Optional character string vector to select scenarios to be
used in the management preparation if not all farmR projects in the SWAT+
model folder should be used (default \code{NULL} all farmR projects are used).}

\item{synonyms}{Optional table to define synonymous op_data1 definitions for
op_types in status quo and a scenario.}

\item{start_year}{Year to start management schedules if not entire scheduled
periods from the farmR projects should be used. \code{start_year} must be
defined together with \code{end_year} (default \code{NULL} all scheduled years are
used). Parameter only used in case of SWATfarmR projects.}

\item{end_year}{Year to end management schedules if not entire scheduled
periods from the farmR projects should be used. \code{end_year} must be defined
together with \code{start_year} (default \code{NULL} all scheduled years are used).
Parameter only used in case of SWATfarmR projects.}

\item{write_path}{Optional path to write the prepared management inputs.}

\item{write_csv_mgts}{Should the management tables with corrected dates be
written as csv files (default is \code{FALSE}). This can be useful to compare
and check the corrected dates of the scenario management tables.}

\item{update_luse_label}{Should the land use labels of the input files be
updated? This is useful if very long names (more than 25 characters) are
used for landuse and management naming. Such long names would be
misinterpreted by SWAT+ simulations. IF \code{TRUE} such labels would be
shortened. If \code{FALSE} (default), this step in the preparation is skipped.}

\item{sync_dates}{If the status quo management and the scenario managements
are mostly similar in their sequences, with only a few measures changed but
different dates for the 'same' operations it can benefit the the
comparability if the dates of those operations are synced between the
status quo and the scenarios (default \code{TRUE}). If this step should be
skipped set \code{FALSE}.}
}
\value{
Writes an '.rds' file into the \code{write_path} which must be used as
the input file to define 'management' related NSWRMs in
\code{measr_project$load_nswrm_definition()}. The file name has the following
structure '<yyyymmdd_hhmm>_management_scenarios.rds'. Optionally also the
management schedules for the scenarios and the status quo are written as
'.csv' files into the folder '<yyyymmdd_hhmm>_management_scenarios'.
}
\description{
Prepare the management input data from SWATfarmR project as inputs for the
definition of management related NSWRMs with
\code{measr_project$load_nswrm_definition()}.
}
