% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_inputs.R
\name{read_cio}
\alias{read_cio}
\title{Read the SWAT+ model's file.cio.}
\usage{
read_cio(file_path)
}
\arguments{
\item{file_path}{Path of the SWAT+ file.cio}
}
\value{
a named list with the lines of file.cio as character vectors.
}
\description{
Read the SWAT+ model's file.cio.
}
\keyword{internal}
