% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_inputs.R
\name{read_def}
\alias{read_def}
\title{Read a SWAT+ .def input file.}
\usage{
read_def(file_path)
}
\arguments{
\item{file_path}{Path of the SWAT+ input file.}
}
\value{
The connecitivity input file as a tibble.
}
\description{
Read a SWAT+ .def input file.
}
\keyword{internal}
