% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_inputs.R
\name{read_dtl_names}
\alias{read_dtl_names}
\title{Read the names of decision table definitions from a SWAT+ dtl input file.}
\usage{
read_dtl_names(file_path)
}
\arguments{
\item{file_path}{Path of the SWAT+ input file.}
}
\value{
A character vector with the names of the defined decision rule sets.
}
\description{
Read the names of decision table definitions from a SWAT+ dtl input file.
}
\keyword{internal}
