% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_management.R
\name{read_farmr}
\alias{read_farmr}
\title{Load a farmR project}
\usage{
read_farmr(file)
}
\arguments{
\item{file}{'.farm' file}
}
\value{
A loaded farmR environment
}
\description{
Load a farmR project
}
\keyword{internal}
