% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write_inputs.R
\name{read_mgt_sch}
\alias{read_mgt_sch}
\title{Read SWAT+ management schedule file.}
\usage{
read_mgt_sch(file_path)
}
\arguments{
\item{file_path}{Path of the SWAT+ management.sch input file.}
}
\value{
The SWAT+ management.sch input file as a tibble.
}
\description{
Read SWAT+ management schedule file.
}
\keyword{internal}
