% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_inputs.R
\name{read_swat_inputs}
\alias{read_swat_inputs}
\title{Read SWAT+ input files and arrange them in a list object}
\usage{
read_swat_inputs(project_path)
}
\arguments{
\item{project_path}{Path to the SWAT project folder on the hard drive
(i.e. txtinout folder).}
}
\value{
A list object with the SWAT+ input files organized in tibbles.
}
\description{
Read SWAT+ input files and arrange them in a list object
}
\keyword{internal}
