% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write_inputs.R
\name{read_tbl2}
\alias{read_tbl2}
\title{Read SWAT+ management schedule file.}
\usage{
read_tbl2(file_path, def_names, par_names, id_num = NULL)
}
\arguments{
\item{file_path}{Path of the SWAT+ management.sch input file.}
}
\value{
The SWAT+ management.sch input file as a tibble.
}
\description{
Read SWAT+ management schedule file.
}
\keyword{internal}
