% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write_inputs.R
\name{read_tbl_n}
\alias{read_tbl_n}
\title{Read SWAT+ management schedule file.}
\usage{
read_tbl_n(
  file_path,
  col_names = NULL,
  n_skip = 1,
  id_col_sel = NULL,
  id_num = NULL
)
}
\arguments{
\item{file_path}{Path of the SWAT+ management.sch input file.}
}
\value{
The SWAT+ management.sch input file as a tibble.
}
\description{
Read SWAT+ management schedule file.
}
\keyword{internal}
