% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implement_wetland.R
\name{remove_connection}
\alias{remove_connection}
\title{Remove connections from a connectivity input table.}
\usage{
remove_connection(obj_con, id_i, obj_typ_rmv = NULL, hyd_typ_rmv = NULL)
}
\arguments{
\item{obj_con}{The object connectivity input table}

\item{id_i}{Object ID for which connections should be removed.}

\item{obj_typ_rmv}{Vector of object types into which the object with \code{id_i}
routes and which should be removed.}

\item{hyd_typ_rmv}{Vector of hydrological flow types into the object with
\code{id_i} routes and which should be removed.}
}
\value{
The updated connectivity input table.
}
\description{
Remove all connections which route either into a certain object type
(\code{obj_typ_rmv}) or which are of a certain hydrological type (\code{hyd_typ_rmv}).
The connections are removed for the entry with the ID \code{id_i}.
}
\keyword{internal}
