% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_landuse_labels.R
\name{remove_consonants}
\alias{remove_consonants}
\title{Remove single consonants of words if they are longer than 12 characters}
\usage{
remove_consonants(tbl)
}
\arguments{
\item{tbl}{Tibble with labels and label lengths}
}
\description{
Remove single consonants of words if they are longer than 12 characters
}
\keyword{internal}
