% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implement_pond.R, R/implement_wetland.R
\name{update_cha_con_pond}
\alias{update_cha_con_pond}
\title{Update the chandeg.con input table.}
\usage{
update_cha_con_pond(cha_con, from_cha_id, res_id)

update_cha_con_pond(cha_con, from_cha_id, res_id)
}
\arguments{
\item{cha_con}{chandeg.con input table}

\item{from_cha_id}{Vector of IDs for the channels which should be routed to
the new reservoir.}

\item{res_id}{New ID of the pond (reservoir) which replaces the land object.}
}
\value{
Updated chandeg.con input table.

Updated chandeg.con input table.
}
\description{
For the channels which should now route into the new reservoir replace all
connections which the channels initially had and add the connection to the
new reservoir.

For the channels which should now route into the new reservoir replace all
connections which the channels initially had and add the connection to the
new reservoir.
}
\keyword{internal}
