% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implement_wetland.R
\name{update_cha_con_wetl}
\alias{update_cha_con_wetl}
\title{Update the chandeg.con input table.}
\usage{
update_cha_con_wetl(cha_con, from_cha_id, hru_id)
}
\arguments{
\item{cha_con}{chandeg.con input table.}

\item{from_cha_id}{Vector of channel IDs which should be routed into the new
wetland.}

\item{hru_id}{HRU IDs to which surface water storage (wetland) will be added.}
}
\value{
Updated chandeg.con table.
}
\description{
In the case channel objects should be rerouted into the new wetland then the
chandeg.con input file must be updated. For the channel which is rerouted
into the wetland all connections are removed and the connection to the
wetland is added.
}
\keyword{internal}
