% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implement_pond.R
\name{update_cio_pond}
\alias{update_cio_pond}
\title{Update the file.cio input table.}
\usage{
update_cio_pond(file_cio)
}
\arguments{
\item{file_cio}{List of file.cio lines.}
}
\value{
Updated file.cio table.
}
\description{
The reservoir input tables are added for the case they were missing, because
no reservoirs were implemented in the initial model setup.
}
\keyword{internal}
