% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_management.R
\name{update_dates}
\alias{update_dates}
\title{Update operation dates for scheduled scenario operations with the dates
of the respective operations in the status quo management schedule.}
\usage{
update_dates(sch_quo, sch_scn, sch_name, synonyms = NULL, max_day_diff = 21)
}
\arguments{
\item{sch_quo}{Operation schedule table for the status quo.}

\item{sch_scn}{Operation schedule table for the scenario case.}

\item{sch_name}{Name of the operation schedule.}

\item{synonyms}{Synonyms input table which defines synonymous operations.}

\item{max_day_diff}{Maximum difference in days between scheduled operations
to consider them as being the same one.}
}
\value{
A date vector with the corrected dates for the scenario schedule.
}
\description{
Update operation dates for scheduled scenario operations with the dates
of the respective operations in the status quo management schedule.
}
\keyword{internal}
