% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implement_pond.R, R/implement_wetland.R
\name{update_hru_con_pond}
\alias{update_hru_con_pond}
\title{Update the hru.con input table.}
\usage{
update_hru_con_pond(hru_con, hru_id)

update_hru_con_pond(hru_con, hru_id)
}
\arguments{
\item{hru_con}{hru.con input table.}

\item{hru_id}{ID of the HRU which is replaced by a pond.}
}
\value{
Updated hru.con input table.

Updated hru.con input table.
}
\description{
To remove HRU from land process calculations the area is set to the minimum
value of 0.00001.

To remove HRU from land process calculations the area is set to the minimum
value of 0.00001.
}
\keyword{internal}
