% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implement_wetland.R
\name{update_hru_hru_wetl}
\alias{update_hru_hru_wetl}
\title{Update the hru-data.hru input table.}
\usage{
update_hru_hru_wetl(hru_hru, lum_i_name, hru_id, hru_id_chr)
}
\arguments{
\item{hru_hru}{hru-data.hru input table.}

\item{lum_i_name}{Name of the land use in landuse.lum which is implemented
for the wetland.}

\item{hru_id}{HRU IDs to which surface water storage (wetland) will be added.}

\item{hru_id_chr}{HRU IDs to which surface water storage (wetland) will be
added. The input is character with leading zeros.}
}
\value{
Updated hru-data.hru input table.
}
\description{
The new landuse pointer to the entry in landuse.lum and the pointer to the
pointer to the now surface storage in wetland.wet are added to the HRU
}
\keyword{internal}
