% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implement_pond.R
\name{update_hyd_res_pond}
\alias{update_hyd_res_pond}
\title{Update the hydrology.res input table.}
\usage{
update_hyd_res_pond(hyd_res, res_id, area)
}
\arguments{
\item{hyd_res}{hydrology.res input table.}

\item{res_id}{New ID of the pond (reservoir) which replaces the land object.}

\item{area}{Area of the replaced land object in ha.}
}
\value{
Updated hydrology.res input table.
}
\description{
Hydrology parameters for the new reservoir are added to hydrology.res.
Most of the parameters are default for all reservoirs. Default the area and
volume parameters are simple functions of the area. The area is the same as
the replaced land object.
}
\keyword{internal}
