% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implement_wetland.R
\name{update_hyd_wet}
\alias{update_hyd_wet}
\title{Update the hydrology.wet input table.}
\usage{
update_hyd_wet(hyd_wet, hyd_wet_i, hru_id_chr)
}
\arguments{
\item{hyd_wet}{hydrology.wet input table.}

\item{hyd_wet_i}{hydrology.wet parameters table for added surface water
storage.}

\item{hru_id_chr}{HRU IDs to which surface water storage (wetland) will be
added. The input is character with leading zeros.}
}
\value{
Updated hydrology.wet input table.
}
\description{
Hydrology parameters for the implemented water wetland water storage are
added to hydrology.hyd. The added parameters are either user defined
(provided with the wetlands definition file) or default if not provided (same
values as set default by SWAT+Editor).
}
\keyword{internal}
