% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implement_nswrm.R
\name{update_implemented_nswrms}
\alias{update_implemented_nswrms}
\title{Update the table implemented_nswrms which tracks the objects (currently HRUs)
in which measures were implemented.}
\usage{
update_implemented_nswrms(
  tbl,
  nswrm_i,
  obj_typ_i,
  obj_id_i,
  obj_typ_upd_i,
  obj_id_upd_i
)
}
\arguments{
\item{tbl}{Existing implemented_nswrms table (if not existing this value is NULL)}

\item{nswrm_i}{String type of the implemented measure.}

\item{obj_typ_i}{String giving the type of spatial objects in which a measure
is implemented (e.g. 'hru').}

\item{obj_id_i}{Vector of integer ID values of the objects in which the
measure is implemented.}

\item{obj_typ_upd_i}{String giving the type of spatial object which replaces
the existing spatial objects with the measure implementation (e.g. 'res' in
case of a pond).}

\item{obj_id_upd_i}{Vector of integer ID values of the new objects which
replaced the initial spatial objects (e.g. res IDs of new reservoirs).}
}
\value{
A tibble with the columns \code{nswrm} for the NSWRM types, \code{obj_typ} with
the type of spatial objects, \code{obj_id} with the IDs of the spatial objects
in which a measure was implemented, \code{obj_typ_new} with the type of the new
spatial objects, and \code{obj_id_new} with the IDs of the new objects.
}
\description{
Update the table implemented_nswrms which tracks the objects (currently HRUs)
in which measures were implemented.
}
\keyword{internal}
