% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_landuse.R
\name{update_landuse}
\alias{update_landuse}
\title{Update the landuse of selected HRUs}
\usage{
update_landuse(
  swat_inputs,
  hru_id,
  type,
  plnt_com = NULL,
  lum_mgt = NULL,
  lum_cn2 = NULL,
  lum_cpr = NULL,
  lum_ovn = NULL
)
}
\arguments{
\item{swat_inputs}{List with SWAT+ input files.}

\item{hru_id}{HRU IDs for which the land use is updated.}

\item{type}{Type of the new land cover. This will be the new lan duse label
in the updated landuse.lum. \code{type} must be a single character string}

\item{plnt_com}{The plant community which is assigned to the land use. A
single character string must be provided. The passed value must be defined
in the 'plant.ini' intput file. If (default) \code{NULL}, then 'plnt_com' will
be set 'null' in the landuse.lum.}

\item{lum_mgt}{The management which is assigned to the land use. A single
character string must be provided. The passed value must be defined in the
'management.sch' intput file. If (default) \code{NULL}, then 'mgt' will be set
'null' in the landuse.lum.}

\item{lum_cn2}{The cn2 parameter set which is assigned to the land use. A
single character string must be provided. The passed value must be defined
in the 'cntabe.lum' intput file. If (default) \code{NULL}, then 'cn2' will be
set 'null' in the landuse.lum.}

\item{lum_cpr}{The conservation practice which is assigned to the land use. A
single character string must be provided. The passed value must be defined
in the 'cons_practice.lum' intput file. If (default) \code{NULL}, then
'cons_prac' will be set 'null' in the landuse.lum.}

\item{lum_ovn}{The Mannings n parametrization which is assigned to the land
use. A single character string must be provided. The passed value must be
defined in the 'ovn_table.lum' intput file. If (default) \code{NULL}, then
'ov_mann' will be set 'null' in the landuse.lum.}
}
\value{
The SWAT+ input tables list with the updated landuse.lum and
hru-data.hru input tables.
}
\description{
The land use ('lu-mgt') in hru-data.hru of the land objects defined by the
\code{hru_id} is overwritten. The new land use label is defined by \code{type}. For
\code{type} a new land use is defined in landuse.lum if no land use with this name
already exists in landuse.lum which has the exact same parametrization as
defined with the additional input arguments. If an identical landuse already
exists, this landuse is assigned to the HRUs. If not a new one is added with
the defined parameters (\code{plnt_com}, \code{lum_mgt}, \code{lum_cn2}, \code{lum_cpr}, and
\code{lum_ovn}).
}
