% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_landuse_labels.R
\name{update_landuse_labels}
\alias{update_landuse_labels}
\title{Update the land use labels which were written by SWATfarmR}
\usage{
update_landuse_labels(project_path)
}
\arguments{
\item{project_path}{Path to the SWAT+ project folder (i.e. TxtInOut).}
}
\value{
Rewrites the input files hru-data.hru, landuse.lum, management.sch,
and plant.ini with shorter labels.
}
\description{
The maximum length for labels in input files which point to entries in other
input files are limited to 16 characters. SWATfarmR writes labels for
landuses, plant communities, and managements by adding index values. They
can result in labels longer than 16 characters. This routine fixes this issue.
}
\keyword{internal}
