% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implement_wetland.R
\name{update_lum_wetl}
\alias{update_lum_wetl}
\title{Update the landuse.lum input table.}
\usage{
update_lum_wetl(luse_lum, hru_hru, lu_mgt_sel, hru_id, hru_id_chr)
}
\arguments{
\item{luse_lum}{landuse.lum input table.}

\item{hru_hru}{hru-data.hru input table.}

\item{lu_mgt_sel}{Selected land use which should be employed for the wetland.
(User defined in the wetlands definition table or NA if not set)}

\item{hru_id}{HRU IDs to which surface water storage (wetland) will be added.}

\item{hru_id_chr}{HRU IDs to which surface water storage (wetland) will be
added. The input is character with leading zeros.}
}
\value{
Updated landuse.lum input table.
}
\description{
By default drainage is removed for HRUs where wetlands are implemented. If
the landuse of a transformed HRU has tile drainage, a new landuse.lum entry
is generated (if it does not already exist). The landuse which is used by the
wetland HRU is always moved to the bottom of the landuse.lum table ( makes it
easier for following functions to identify the employed land use).
}
\keyword{internal}
