% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implement_nswrm.R
\name{update_luse_by_type}
\alias{update_luse_by_type}
\title{Title}
\usage{
update_luse_by_type(type_i, nswrm_loc_sel, nswrm_defs, swat_inputs)
}
\arguments{
\item{swat_inputs}{}
}
\description{
Title
}
\keyword{internal}
