% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_management.R
\name{update_management}
\alias{update_management}
\title{Update the management inputs for selected HRUs}
\usage{
update_management(swat_inputs, hru_id, mgt_def)
}
\arguments{
\item{swat_inputs}{List with SWAT+ input files.}

\item{hru_id}{HRU IDs for which the land use is updated.}

\item{mgt_def}{List of management input files which define the management
scenario.}
}
\value{
The SWAT+ input tables list with the updated hru-data.hru,
landuse.lum, management.sch, and plant.ini input tables.
}
\description{
Update the management inputs for selected HRUs
}
\keyword{internal}
