% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_nswrm_definition.R
\name{update_nswrm_lookup}
\alias{update_nswrm_lookup}
\title{Add entries for NSWRMs in the NSWRM lookup table}
\usage{
update_nswrm_lookup(nswrm_lookup, type, nswrm, overwrite)
}
\arguments{
\item{nswrm_lookup}{NSWRM lookup table}

\item{type}{NSWRM type in which the NSWRMs are grouped based on their
implementation.}

\item{nswrm}{Character vector of NSWRM labels.}

\item{overwrite}{If an NSWRM definition is overwritten, then the lookup is
also replaced}
}
\value{
The NSWRM lookup tibble with the newly added entries.
}
\description{
Add entries for NSWRMs in the NSWRM lookup table
}
\keyword{internal}
