% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implement_pond.R
\name{update_obj_cnt_pond}
\alias{update_obj_cnt_pond}
\title{Update the object.cnt input table.}
\usage{
update_obj_cnt_pond(obj_cnt)
}
\arguments{
\item{obj_cnt}{object.cnt input table.}
}
\value{
Updated object.cnt input table.
}
\description{
The total object count and the count of reservoirs are increased by one.
}
\keyword{internal}
