% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implement_pond.R, R/implement_wetland.R
\name{update_res_con_pond}
\alias{update_res_con_pond}
\title{Update the reservoir.con input table.}
\usage{
update_res_con_pond(res_con, rtu_con, hru_id, to_cha_id, res_id)

update_res_con_pond(res_con, rtu_con, hru_id, to_cha_id, res_id)
}
\arguments{
\item{res_con}{reservoir.con input table}

\item{rtu_con}{rout_unit.con input table.}

\item{hru_id}{ID of the HRU which is replaced by a pond.}

\item{to_cha_id}{ID of channel to which water from reservoir is routed.}

\item{res_id}{New ID of the pond (reservoir) which replaces the land object.}
}
\value{
Updated reservoir.con input table with new line for the new
reservoir.

Updated reservoir.con input table with new line for the new
reservoir.
}
\description{
Take the properties of the routing unit (area, lat, lon, elev...) and use
them for the newly generated reservoir. Add those properties as a new line
in reservoir.con and add the connection to the channel.

Take the properties of the routing unit (area, lat, lon, elev...) and use
them for the newly generated reservoir. Add those properties as a new line
in reservoir.con and add the connection to the channel.
}
\keyword{internal}
