% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implement_pond.R
\name{update_res_res_pond}
\alias{update_res_res_pond}
\title{Update the reservoir.res input table.}
\usage{
update_res_res_pond(res_res)
}
\arguments{
\item{res_res}{reservoir.res input table}
}
\value{
Updated reservoir.res input table.
}
\description{
Add a new line in the reservoir.res table with the ID and name of the new
reservoir. Only the hydrology column \code{hyd} and the ID are updated. The other
pointers are set to the default entries (currently the case for all
reservoirs).
}
\keyword{internal}
