% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implement_pond.R, R/implement_wetland.R
\name{update_rtu_con_pond}
\alias{update_rtu_con_pond}
\title{Update the rout_unit.con input table.}
\usage{
update_rtu_con_pond(rtu_con, rtu_con_chg, hru_id, res_id)

update_rtu_con_pond(rtu_con, rtu_con_chg, hru_id, res_id)
}
\arguments{
\item{rtu_con}{rout_unit.con input table.}

\item{rtu_con_chg}{Lookup table for which connectivities must be changed.}

\item{hru_id}{ID of the HRU which is replaced by a pond.}

\item{res_id}{New ID of the pond (reservoir) which replaces the land object.}
}
\value{
Updated rout_unit.con table.

Updated rout_unit.con table.
}
\description{
All connections of land objects which initially routed into the replaced land
object (HRU) are replaced by a connection into the newly generated reservoir.
For all connections of the replaced land object the fractions are set to 0.
To elimiate the effect of e.g. ET from that object its area is set to the
minimum value (0.00001).

All connections of land objects which initially routed into the replaced land
object (HRU) are replaced by a connection into the newly generated reservoir.
For all connections of the replaced land object the fractions are set to 0.
To elimiate the effect of e.g. ET from that object its area is set to the
minimum value (0.00001).
}
\keyword{internal}
