% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implement_wetland.R
\name{update_rtu_con_wetl}
\alias{update_rtu_con_wetl}
\title{Update the rout_unit.con input table.}
\usage{
update_rtu_con_wetl(rtu_con, to_cha_id, has_drn, hru_id)
}
\arguments{
\item{rtu_con}{rout_unit.con input table.}

\item{to_cha_id}{ID of the channel into which all water should be routed.}

\item{has_drn}{Boolean. Does the HRU have tile drainage or not?}

\item{hru_id}{HRU IDs to which surface water storage (wetland) will be added.}
}
\value{
Updated rout_unit.con table.
}
\description{
If the new wetland should route directly into a channel instead of routing
into its neighboring spatial object, or if the HRU has tile drainage which
should be removed then the rout_unit.con is updated accordingly. In the case
of routing into a channel all connections (except aquifer recharge), are
removed for the wetland HRU and are replaced by a connection to the channel
with the ID \code{to_cha_id}. In the case of tile drainage only the tile flow
connectivity is removed.
}
\keyword{internal}
