% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implement_wetland.R
\name{update_wet_wet}
\alias{update_wet_wet}
\title{Update the wetland.wet input table.}
\usage{
update_wet_wet(wet_wet, wet_wet_i, hru_id_chr)
}
\arguments{
\item{wet_wet}{wetland.wet input table}

\item{wet_wet_i}{Table with rel, sed, and nut pointers for added pond.}

\item{hru_id_chr}{HRU IDs to which surface water storage (wetland) will be
added. The input is character with leading zeros.}
}
\value{
Updated wetland.wet input table.
}
\description{
Add a new line in the wetland.wet table. The name of the added entry is 'wet'
and the HRU ID. The wetland is initialized with the default 'initwet1'. The
pointer to the right entry in hydrology.wet is initialized ('hydwet' + HRU
ID). The (user defined) pointers to release, sediments, and nutrients are
added.
}
\keyword{internal}
