% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_inputs.R
\name{write_cio}
\alias{write_cio}
\title{Write SWAT+ file.cio file.}
\usage{
write_cio(file_cio, file_path, fmt)
}
\arguments{
\item{file_cio}{SWAT file.cio.}

\item{file_path}{Write path of the SWAT+ input file.}
}
\value{
Writes the file.cio as a text file in the file path.
}
\description{
Write SWAT+ file.cio file.
}
\keyword{internal}
