% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_inputs.R
\name{write_def}
\alias{write_def}
\title{Write SWAT+ .def file.}
\usage{
write_def(tbl, file_path, fmt)
}
\arguments{
\item{tbl}{SWAT+ .def table.}

\item{file_path}{Write path of the SWAT+ input file.}

\item{fmt}{Character vector of format strings to define the print format of
each table column.}
}
\value{
Writes the .def table as a text file in the file path.
}
\description{
Write SWAT+ .def file.
}
\keyword{internal}
