% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_management.R
\name{write_farmr_ops}
\alias{write_farmr_ops}
\title{Write scheduled opeartions from farmR project into project folder}
\usage{
write_farmr_ops(farmr_project, start_year, end_year)
}
\arguments{
\item{farmr_project}{SWATfarmR project environment}

\item{start_year}{Start year}

\item{end_year}{End year}
}
\value{
Writes the SWAT+ input files which are relevant for management
scheduling into the project path
}
\description{
Write scheduled opeartions from farmR project into project folder
}
\keyword{internal}
