% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_management.R
\name{write_operation_v4}
\alias{write_operation_v4}
\title{Write the mgt schedules into the respective files in the TxtInOut folder}
\usage{
write_operation_v4(data, start_year, end_year)
}
\arguments{
\item{data}{List with all data stored in the farmR object.}

\item{start_year}{Numeric. Defines the start year for which to write operations.}

\item{end_year}{Numeric. Defines the last year for which to write operations.}
}
\description{
Write the mgt schedules into the respective files in the TxtInOut folder
}
\keyword{internal}
