% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_inputs.R
\name{write_swat_inputs}
\alias{write_swat_inputs}
\title{Write the SWAT+ input tables from the \code{swat_input} list object which were
updated by implementing NSWRMs to SWAT+ input text files.}
\usage{
write_swat_inputs(swat_inputs, project_path)
}
\arguments{
\item{swat_inputs}{List of SWAT+ input tables.}

\item{project_path}{Path to the SWAT project folder on the hard drive (i.e.
txtinout folder).}
}
\value{
Writes SWAT+ text input files into the \code{project_path}.
}
\description{
Write the SWAT+ input tables from the \code{swat_input} list object which were
updated by implementing NSWRMs to SWAT+ input text files.
}
