% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_management.R
\name{add_op_date}
\alias{add_op_date}
\title{Title}
\usage{
add_op_date(tbl, start_year)
}
\arguments{
\item{tbl}{Operation schedule table}

\item{start_year}{Start year for the operation schedule}
}
\value{
The table with an added date column
}
\description{
Title
}
\keyword{internal}
