% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_nswrm_definition.R
\name{load_nswrm_def}
\alias{load_nswrm_def}
\title{Load a definition table for an NSWRM type from a '.csv' file into the
\code{measr_project}.}
\usage{
load_nswrm_def(file_path, type, nswrm_defs, swat_inputs, overwrite)
}
\arguments{
\item{file_path}{Path to the '.csv' or '.rds' definition file.}

\item{type}{Type of the NSWRM which is defined by this input file. The
type must be one of the options \code{'land_use'}, \code{'pond'}.}

\item{nswrm_defs}{List with already loaded NSWRM definition tables}

\item{swat_inputs}{List with SWAT+ input files.}

\item{overwrite}{Overwrite existing definition table? If \code{TRUE} existing
definition table can be overwritten.}
}
\value{
The list with NSWRM definition tables with the new added entry.
}
\description{
Load a definition table for an NSWRM type from a '.csv' file into the
\code{measr_project}.
}
\details{
The different \code{type} options represent different groups of NSWRMs. The
different options group the following measures:
\itemize{
\item \code{'land_use'}: A land use measure definition table can include the
definitions for all NSWRMs which are represented by a land use
change. The lookup table can include the settings for the NSWRM codes
\strong{buffer}, \strong{edgefilter}, \strong{hedge}, \strong{grassslope}, \strong{grassrchrg}
, and \strong{afforest}. The \code{land_use} table must provide the columns
\code{type}, \code{lum_plnt}, \code{lum_mgt}, \code{lum_cn2}, \code{lum_cpr}, and
\code{lum_ovn}.
\item \code{'pond'}: A pond definition table includes all definitions for pond
locations. The \code{pond} table must provide the columns \code{name},
\code{to_cha_id}, and \code{from_cha_id}.
}
}
\keyword{internal}
