% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measr_object.R
\docType{class}
\name{measr_project}
\alias{measr_project}
\alias{new}
\title{\code{measr_project} is an \code{R6Class} to include NSWRMs in a SWAT+ model setup.}
\usage{
new_measr(project_name, project_path)
}
\value{
Returns a \code{swat_measr} object in the \code{R} environment and saves the
project as a single data file with the name '<project_name>.measr' in
\code{project_path}.
}
\description{
A \code{measr_project} is an \code{R} object which stores all data (SWAT+
model input files, NSWRM definitions, potential measure locations) and the
corresponding functions to implement NSWRMs in the text input files of a
SWAT+ project.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{.data}}{List which stores the all input data that is required to
implement NSWRMs in the SWAT+ project which is located in
\code{project_path}. \code{.data} has the following elements:
\itemize{
\item \code{.data$model_setup}: stores all SWAT+ input files that were loaded
from the \code{project_path}. Two versions of the input files are saved in
\code{.data$model_setup}
\itemize{
\item \code{.data$model_setup$original_inputs} are the unchanged input files
which are required to reset to an initial condition.
\item \code{.data$model_setup$modified_inputs} are the input files which are
modified when NSWRMs are implemented.
}
}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-measr_project-new}{\code{measr_project$new()}}
\item \href{#method-measr_project-load_nswrm_definition}{\code{measr_project$load_nswrm_definition()}}
\item \href{#method-measr_project-save}{\code{measr_project$save()}}
\item \href{#method-measr_project-reset}{\code{measr_project$reset()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-measr_project-new"></a>}}
\if{latex}{\out{\hypertarget{method-measr_project-new}{}}}
\subsection{Method \code{new()}}{
Create a new SWATmeasR project. To start a new
\code{meas_project} use the function \code{new_measr(project_name, project_path)}
instead of the class function \code{measr_project$new()}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{measr_project$new(project_name, project_path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{project_name}}{Name of the \code{measr_project}. This name is assigned to
the object in the R environment and is also the name of the saved
object in the SWAT+ project folder.}

\item{\code{project_path}}{Path to the SWAT project folder on the hard drive
(i.e. txtinout folder).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-measr_project-load_nswrm_definition"></a>}}
\if{latex}{\out{\hypertarget{method-measr_project-load_nswrm_definition}{}}}
\subsection{Method \code{load_nswrm_definition()}}{
Load a definition table for an NSWRM type from a '.csv' file into the
\code{measr_project}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{measr_project$load_nswrm_definition(file_path, type, overwrite)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file_path}}{Path to the '.csv' definition file.}

\item{\code{type}}{Type of the NSWRM which is defined by this input file. The
type must be one of the options \code{'land_use'}, \code{'pond'}.}

\item{\code{overwrite}}{Overwrite existing definition table? Default is \code{FALSE}.
If \code{TRUE} existing definition table can be overwritten.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
The different \code{type} options represent different groups of NSWRMs. The
different options group the following measures:
\itemize{
\item \code{'land_use'}: A land use measure definition table can include the
definitions for all NSWRMs which are represented by a land use
change. The lookup table can include the settings for the NSWRM codes
\strong{buffer}, \strong{edgefilter}, \strong{hedge}, \strong{grassslope}, \strong{grassrchrg}
, and \strong{afforest}. The \code{land_use} table must provide the columns
\code{nswrm}, \code{lum_plnt}, \code{lum_mgt}, \code{lum_cn2}, \code{lum_cpr}, and
\code{lum_ovn}.
\item \code{'pond'}: A pond definition table includes all definitions for pond
locations. The \code{pond} table must provide the columns \code{name},
\code{to_cha_id}, and \code{from_cha_id}.
}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-measr_project-save"></a>}}
\if{latex}{\out{\hypertarget{method-measr_project-save}{}}}
\subsection{Method \code{save()}}{
Save the SWATmeasR project in the SWAT+ project folder.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{measr_project$save()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-measr_project-reset"></a>}}
\if{latex}{\out{\hypertarget{method-measr_project-reset}{}}}
\subsection{Method \code{reset()}}{
Reset the changes applied to the SWAT+ input files.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{measr_project$reset()}\if{html}{\out{</div>}}
}

}
}
