% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_management.R
\name{prepare_management_scenario_inputs}
\alias{prepare_management_scenario_inputs}
\title{Prepare the management input data from SWATfarmR project as inputs
for the definition of management related NSWRMs with
\code{measr_project$load_nswrm_definition()}.}
\usage{
prepare_management_scenario_inputs(
  project_path,
  status_quo,
  scenarios = NULL,
  synonyms = NULL,
  start_year = NULL,
  end_year = NULL,
  write_path = project_path,
  write_csv_mgts = FALSE
)
}
\arguments{
\item{project_path}{Path to the SWAT project folder on the hard drive
(i.e. txtinout folder).}

\item{status_quo}{Name of the SWATfarmR project which provides the status
quo.}

\item{synonyms}{Optional table to define synonymous op_data1 definitions for
op_types in status quo and a scenario.}

\item{write_path}{Optional path to write the prepared management inputs.}

\item{write_csv_mgts}{Should the management tables with corrected dates be
written as csv files (default is \code{FALSE}). This can be useful to compare
and check the corrected dates of the scenario management tables.}
}
\value{
Writes an '.rds' file into the \code{write_path} which must be used as
the input file to define 'management' related NSWRMs in
\code{measr_project$load_nswrm_definition()}. The file name has the following
structure '<yyyymmdd_hhmm>_management_scenarios.rds'. Optionally also
the management schedules for the scenarios and the status quo are written
as '.csv' files into the folder '<yyyymmdd_hhmm>_management_scenarios'.
}
\description{
Prepare the management input data from SWATfarmR project as inputs
for the definition of management related NSWRMs with
\code{measr_project$load_nswrm_definition()}.
}
