% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_inputs.R
\name{read_tbl}
\alias{read_tbl}
\title{Read a SWAT+ input file which has a tabular structure.}
\usage{
read_tbl(file_path, col_names = NULL, n_skip = 1)
}
\arguments{
\item{file_path}{Path of the SWAT+ input file.}

\item{col_names}{(optional) Column names vector.}

\item{n_skip}{Number of header rows to skip. Default is 1.}
}
\value{
The SWAT+ input file as a tibble.
}
\description{
Read a SWAT+ input file which has a tabular structure.
}
\keyword{internal}
