% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implement_wetland.R
\name{update_wet_wet}
\alias{update_wet_wet}
\title{Update the reservoir.res input table.}
\usage{
update_wet_wet(hyd_wet, hyd_res_pnd, hru_id, area)

update_wet_wet(hyd_wet, hyd_res_pnd, hru_id, area)
}
\arguments{
\item{hyd_res_pnd}{hydrology.res parameters table for added pond.}

\item{hru_id}{ID of the HRU which is replaced by a pond.}

\item{area}{Area of the replaced land object in ha.}

\item{res_res}{reservoir.res input table}

\item{res_res_pnd}{table with rel, sed, and nut pointers for added pond.}

\item{hyd_res}{hydrology.res input table.}
}
\value{
Updated reservoir.res input table.

Updated hydrology.res input table.
}
\description{
Add a new line in the reservoir.res table with the ID and name of the new
reservoir. Only the hydrology column \code{hyd} and the ID are updated. The other
pointers are set to the default entries (currently the case for all
reservoirs).

Hydrology parameters for the new reservoir are added to hydrology.res.
Most of the parameters are default for all reservoirs. Default the area and
volume parameters are simple functions of the area. The area is the same as
the replaced land object.
}
\keyword{internal}
